test: move runtime test to src/node
