+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/assets.js
