+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/b.ts
