+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/c.ts
