+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/circular/circular-index.js
