+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/cjs-external-existing.js
