+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/esm-external-existing.js
