+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/esm-external-non-existing.js
