+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/esm-external/index.mjs
