+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/has-error.js
