+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/simple.js
