+++ packages/vite/src/node/ssr/runtime/__tests__/fixtures/throws-error-method.ts
