+++ packages/vite/src/node/ssr/runtime/__tests__/package.json
